+++ src/pages/LabSheet.jsx
@@ -4,6 +4,9 @@ import Header from "../components/Header";
 import Sidenav from "../components/Sidenav";
 import { Box, Flex } from "@chakra-ui/react";
 // import { EllipsisVerticalIcon } from "@heroicons/react/24/outline";
+import { useParams } from "react-router-dom";
+import { supabase } from "../config/supabase";
+import { useEffect, useState } from "react";
 
 import {
   useSession,
@@ -15,6 +18,25 @@ function LabSheets() {
   // const { isLoading } = useSessionContext();
   const session = useSession();
 
+  const organization_id = useParams().organization_id;
+
+  const [organization, setOrganization] = useState("");
+
+  useEffect(() => {
+    const fetchOrganization = async () => {
+      const { data, error } = await supabase
+        .from("Organizations")
+        .select("*")
+        .eq("organization_id", organization_id);
+
+      if (!error) {
+        setOrganization(data[0]);
+      }
+    };
+
+    fetchOrganization();
+  }, [organization_id]);
+
   // if (isLoading) {
   //     return <></>;
   // }
@@ -24,7 +46,7 @@ function LabSheets() {
       <Header />
       {session ? (
         <Box flex={1} display={"flex"} flexDirection={"row"} zIndex={1}>
-          <Sidenav />
+          <Sidenav organization={organization}/>
           <Flex flex={1} flexDirection={"column"} alignItems="center">
             <iframe
               title="google-drive"
